+++ src/dfx-core/src/extension/manifest/mod.rs
@@ -1,14 +1,15 @@
 //! Directory contains code that parses the .json files.
 
 pub mod compatibility_matrix;
+pub mod dependencies;
+pub mod extension;
+
 /// `compatibility.json` is a file describing the compatibility
 /// matrix between extensions versions and the dfx version.
 pub use compatibility_matrix::ExtensionCompatibilityMatrix;
-/// URL to the `compatibility.json` file.
-pub use compatibility_matrix::COMMON_EXTENSIONS_MANIFEST_LOCATION;
 
-pub mod extension;
-/// `manifest.json` is a file describing the extension.
+/// A file that lists the dependencies of all versions of an extension.
+pub use dependencies::ExtensionDependencies;
+
+/// A file that describes an extension.
 pub use extension::ExtensionManifest;
-/// File name for the file describing the extension.
-pub use extension::MANIFEST_FILE_NAME;