+++ src/dfx/src/commands/schema.rs
@@ -2,6 +2,7 @@ use crate::lib::{error::DfxResult, metadata::dfx::DfxMetadata};
 use anyhow::Context;
 use clap::{Parser, ValueEnum};
 use dfx_core::config::model::dfinity::{ConfigInterface, TopLevelConfigNetworks};
+use dfx_core::extension::manifest::{ExtensionDependencies, ExtensionManifest};
 use schemars::schema_for;
 use std::path::PathBuf;
 
@@ -10,6 +11,7 @@ enum ForFile {
     Dfx,
     Networks,
     DfxMetadata,
+    ExtensionDependencies,
     ExtensionManifest,
 }
 
@@ -28,9 +30,8 @@ pub fn exec(opts: SchemaOpts) -> DfxResult {
     let schema = match opts.r#for {
         Some(ForFile::Networks) => schema_for!(TopLevelConfigNetworks),
         Some(ForFile::DfxMetadata) => schema_for!(DfxMetadata),
-        Some(ForFile::ExtensionManifest) => {
-            schema_for!(dfx_core::extension::manifest::ExtensionManifest)
-        }
+        Some(ForFile::ExtensionDependencies) => schema_for!(ExtensionDependencies),
+        Some(ForFile::ExtensionManifest) => schema_for!(ExtensionManifest),
         _ => schema_for!(ConfigInterface),
     };
     let nice_schema =